requirements.txt
# dependencies list
# pandas
# requests
# twilio
# python-dotenv
# schedule
# scikit-learn
# streamlit

streamlit==1.37.0
pandas==2.2.3
scikit-learn==1.5.2
twilio==9.8.6
python-dotenv==1.2.1
schedule==1.2.2
requests==2.32.5
